requirements.txt
moviepy==1.0.3
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.4
gspread==5.12.0
Pillow==10.1.0
numpy==1.24.3
google-auth==2.23.4
PyDrive2==1.17.0
schedule==1.2.0
python-dateutil==2.8.2
requests==2.31.0 
python-dotenv